+++ flopp-kernel/src/main/java/com/github/kjetilv/flopp/kernel/bits/AbstractBitwiseCsvLineSplitter.java
@@ -143,7 +143,7 @@ abstract sealed class AbstractBitwiseCsvLineSplitter extends AbstractBitwiseLine
 
     @Override
     public final int hashCode() {
-        return LineSegments.hashCode(this);
+        return (int) LineSegments.hashCode(this);
     }
 
     @Override
